add authorization

+++ internal/database/tasks.go
@@ -17,15 +17,16 @@ func NewTaskStore(db *sqlx.DB) *TaskStore {
 	return &TaskStore{db: db}
 }
 
-func (s *TaskStore) GetAll() ([]models.Task, error) {
+func (s *TaskStore) GetAll(ownerID int) ([]models.Task, error) {
 	var tasks []models.Task
 
 	query := `
 SELECT * 
 FROM tasks 
-order by created_at desc;`
+WHERE owner_id = $1
+ORDER BY created_at DESC;`
 
-	err := s.db.Select(&tasks, query)
+	err := s.db.Select(&tasks, query, ownerID)
 	if err != nil {
 		return nil, err
 	}
@@ -33,18 +34,17 @@ order by created_at desc;`
 	return tasks, nil
 }
 
-func (s *TaskStore) GetByID(id int) (*models.Task, error) {
+func (s *TaskStore) GetByID(id, ownerID int) (*models.Task, error) {
 	var task models.Task
 
 	query := `
 SELECT * 
 FROM tasks 
-WHERE id = $1;`
-	err := s.db.Get(&task, query, id)
+WHERE id = $1 AND owner_id = $2;`
 
-	// когда запрос не вернул ни одной строки
+	err := s.db.Get(&task, query, id, ownerID)
 	if err == sql.ErrNoRows {
-		return nil, fmt.Errorf(`task with id %d not found`, id)
+		return nil, fmt.Errorf("task with id %d not found", id)
 	}
 
 	if err != nil {
@@ -54,23 +54,23 @@ WHERE id = $1;`
 	return &task, nil
 }
 
-func (s *TaskStore) Create(input models.CreateTaskInput) (*models.Task, error) {
+func (s *TaskStore) Create(input models.CreateTaskInput, ownerID int) (*models.Task, error) {
 	var task models.Task
 	query := `
-INSERT INTO tasks (title, description, completed, created_at, updated_at) 
-VALUES ($1, $2, $3, $4, $5)
+INSERT INTO tasks (title, description, completed, owner_id, created_at, updated_at) 
+VALUES ($1, $2, $3, $4, $5, $6)
 RETURNING *;`
 
 	now := time.Now()
-	err := s.db.QueryRowx(query, input.Title, input.Description, now, now).StructScan(&task)
+	err := s.db.QueryRowx(query, input.Title, input.Description, input.Completed, ownerID, now, now).StructScan(&task)
 	if err != nil {
 		return nil, err
 	}
 	return &task, nil
 }
 
-func (s *TaskStore) Update(id int, input models.UpdateTaskInput) (*models.Task, error) {
-	task, err := s.GetByID(id)
+func (s *TaskStore) Update(id, ownerID int, input models.UpdateTaskInput) (*models.Task, error) {
+	task, err := s.GetByID(id, ownerID)
 	if err != nil {
 		return nil, err
 	}
@@ -81,7 +81,6 @@ func (s *TaskStore) Update(id int, input models.UpdateTaskInput) (*models.Task, 
 	if input.Description != nil {
 		task.Description = *input.Description
 	}
-
 	if input.Completed != nil {
 		task.Completed = *input.Completed
 	}
@@ -90,12 +89,12 @@ func (s *TaskStore) Update(id int, input models.UpdateTaskInput) (*models.Task, 
 
 	query := `
 UPDATE tasks
-SET title = $1, description = $2, completed = $3, updated_at = $4 WHERE id = $5
+SET title = $1, description = $2, completed = $3, updated_at = $4 
+WHERE id = $5 AND owner_id = $6
 RETURNING *;`
 
 	var updatedTask models.Task
-
-	err = s.db.QueryRowx(query, task.Title, task.Description, task.Completed, task.UpdatedAt, id).StructScan(&updatedTask)
+	err = s.db.QueryRowx(query, task.Title, task.Description, task.Completed, task.UpdatedAt, id, ownerID).StructScan(&updatedTask)
 	if err != nil {
 		return nil, err
 	}
@@ -103,10 +102,10 @@ RETURNING *;`
 	return &updatedTask, nil
 }
 
-func (s *TaskStore) Delete(id int) error {
-	query := `DELETE FROM tasks WHERE id = $1;`
+func (s *TaskStore) Delete(id, ownerID int) error {
+	query := `DELETE FROM tasks WHERE id = $1 AND owner_id = $2;`
 
-	result, err := s.db.Exec(query, id)
+	result, err := s.db.Exec(query, id, ownerID)
 	if err != nil {
 		return err
 	}
@@ -116,7 +115,7 @@ func (s *TaskStore) Delete(id int) error {
 		return err
 	}
 	if rows == 0 {
-		return fmt.Errorf(`task with id %d not found`, id)
+		return fmt.Errorf("task with id %d not found", id)
 	}
 	return nil
 }
